+++ CMakeLists.txt
@@ -9,7 +9,8 @@ add_executable(smart_doorlock thread/main.c
         brightness/cds.c
         ultrasonic/ultrasonic.c
         oled/oled.c
-        servo/lock.c)
+        servo/lock.c
+        tone/music.c)
 
 add_executable(tone tone/tone.c)
 add_executable(stepper stepper/stepper.c)
