+++ tone/music.h
@@ -1,0 +1,13 @@
+//
+// Created by dfkdream on 22. 12. 9.
+//
+
+#ifndef SMART_DOORLOCK_MUSIC_H
+#define SMART_DOORLOCK_MUSIC_H
+
+#define MUSIC_DOOR_OPEN 0
+#define MUSIC_DOOR_CLOSE 1
+
+int music(int stat);
+
+#endif //SMART_DOORLOCK_MUSIC_H
